fix(feature/contact: location services
